pyproject.toml
[project]
name = "research-evaluation-pipeline"
version = "0.1.0"
description = "A framework for automated assessment and diagnostic analysis of research papers using LLMs"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.70.0",
    "keyring>=25.7.0",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.3",
    "respx>=0.23.1",
    "tomli>=2.4.1",
    "tomli-w>=1.2.0",
    "typer>=0.24.2",
]

[project.scripts]
rrp = "research_evaluation_pipeline.cli:app"

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
[tool.ruff.format]
skip-magic-trailing-comma = true

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
